requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-pptx>=0.6.21
youtube-transcript-api>=0.6.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
plotly>=5.17.0
pathlib>=1.0.1
typing-extensions>=4.0.0